refactor(packaging): use Yarn 2 link protocol for "src" package
This simplifies module name resolution situation a bit, and removes the
need for "moduleNameMapper" in the Jest config.

The situation might be further improved by using Yarn 2's automatic
self-reference support through the "registree-core" package name (which
would remove the need for the "src" link), but that's not automatically
supported by TypeScript or Jest yet, so I'm holding off on that.

Yarn 2 documentation:

https://yarnpkg.com/features/protocols#why-is-the-link-protocol-recommended-over-aliases-for-path-mapping